Package refactors, changed naming conventions in DTO classes
