+++ src/main/java/com/project/trackfit/aws/ImageService.java
@@ -4,7 +4,7 @@ import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.amazonaws.services.s3.model.S3Object;
-import com.project.trackfit.customer.dto.Customer;
+import com.project.trackfit.customer.entity.Customer;
 import com.project.trackfit.customer.service.CustomerService;
 import com.project.trackfit.media.Media;
 import com.project.trackfit.media.MediaRepository;
