+++ src/main/java/com/project/trackfit/customer/repository/CustomerRepository.java
@@ -1,6 +1,6 @@
 package com.project.trackfit.customer.repository;
 
-import com.project.trackfit.customer.dto.Customer;
+import com.project.trackfit.customer.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
