+++ src/main/java/com/project/trackfit/measurements/MeasurementsService.java
@@ -2,7 +2,7 @@ package com.project.trackfit.measurements;
 
 import com.project.trackfit.core.exception.MeasurementNotFoundException;
 import com.project.trackfit.core.exception.ResourceNotFoundException;
-import com.project.trackfit.customer.dto.Customer;
+import com.project.trackfit.customer.entity.Customer;
 import com.project.trackfit.customer.repository.CustomerRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
