+++ src/main/java/com/project/trackfit/steps/DailyStepsRetrieveRequestMapper.java
@@ -1,6 +1,6 @@
 package com.project.trackfit.steps;
 
-import com.project.trackfit.customer.dto.Customer;
+import com.project.trackfit.customer.entity.Customer;
 import org.springframework.stereotype.Service;
 
 import java.util.function.Function;
