+++ src/main/java/com/project/trackfit/trainer/repository/PersonalTrainerRepository.java
@@ -1,6 +1,6 @@
 package com.project.trackfit.trainer.repository;
 
-import com.project.trackfit.trainer.dto.PersonalTrainer;
+import com.project.trackfit.trainer.entity.PersonalTrainer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
