+++ src/main/java/com/project/trackfit/trainer/service/PersonalTrainerService.java
@@ -1,8 +1,8 @@
 package com.project.trackfit.trainer.service;
 
-import com.project.trackfit.trainer.dto.PersonalTrainer;
+import com.project.trackfit.trainer.entity.PersonalTrainer;
 import com.project.trackfit.trainer.repository.PersonalTrainerRepository;
-import com.project.trackfit.user.dto.ApplicationUser;
+import com.project.trackfit.user.entity.ApplicationUser;
 import com.project.trackfit.core.exception.ResourceNotFoundException;
 
 import org.springframework.stereotype.Service;
