+++ src/test/java/com/project/trackfit/aws/ImageServiceTests.java
@@ -3,7 +3,7 @@ package com.project.trackfit.aws;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.PutObjectRequest;
 import com.amazonaws.services.s3.model.S3Object;
-import com.project.trackfit.customer.dto.Customer;
+import com.project.trackfit.customer.entity.Customer;
 import com.project.trackfit.customer.service.CustomerService;
 import com.project.trackfit.media.Media;
 import com.project.trackfit.media.MediaRepository;
