+++ src/test/java/com/project/trackfit/user/service/ApplicationUserServiceTests.java
@@ -5,7 +5,7 @@ import com.project.trackfit.core.exception.EmailAlreadyTakenException;
 import com.project.trackfit.customer.service.ICustomerService;
 import com.project.trackfit.trainer.service.IPersonalTrainerService;
 import com.project.trackfit.user.component.PasswordCreation;
-import com.project.trackfit.user.dto.ApplicationUser;
+import com.project.trackfit.user.entity.ApplicationUser;
 import com.project.trackfit.user.dto.CreateUserRequest;
 import com.project.trackfit.user.repository.ApplicationUserRepo;
 import org.junit.jupiter.api.DisplayName;
